fix generic input formatsand derived class to use derived objects
