+++ genomix/genomix-data/src/test/java/edu/uci/ics/genomix/data/test/EdgeMapTest.java
@@ -24,7 +24,7 @@ import edu.uci.ics.genomix.type.ReadHeadSet;
 import edu.uci.ics.genomix.type.ReadHeadInfo;
 import edu.uci.ics.genomix.type.ReadIdSet;
 import edu.uci.ics.genomix.type.VKmer;
-import edu.uci.ics.genomix.type.Node.EDGETYPE;
+import edu.uci.ics.genomix.type.EDGETYPE;
 
 public class EdgeMapTest {
