+++ genomix/genomix-hadoop/src/main/java/edu/uci/ics/genomix/hadoop/graph/GraphStatistics.java
@@ -36,7 +36,7 @@ import org.jfree.data.xy.XYSeriesCollection;
 import edu.uci.ics.genomix.config.GenomixJobConf;
 import edu.uci.ics.genomix.type.Node;
 import edu.uci.ics.genomix.type.Node.DIR;
-import edu.uci.ics.genomix.type.Node.EDGETYPE;
+import edu.uci.ics.genomix.type.EDGETYPE;
 import edu.uci.ics.genomix.type.ReadIdSet;
 import edu.uci.ics.genomix.type.VKmer;
